Comment all single process per user locking until indexing can be done without launching a separate mikidown window. Also, force the config files to be ini to make it easier to locate for users.

+++ mikidown/__init__.py
@@ -2,6 +2,7 @@ import argparse
 import os
 import re
 import sys
+#import signal
 
 from PyQt4.QtCore import QSettings
 from PyQt4.QtGui import QApplication, QIcon, QMessageBox
@@ -17,17 +18,6 @@ from .sandbox import Sandbox
 sys.path.append(os.path.dirname(__file__))
 
 # http://code.activestate.com/recipes/578453-python-single-instance-cross-platform/
-def set_exit_handler(func):
-    if os.name == "nt":
-        try:
-            import win32api
-            win32api.SetConsoleCtrlHandler(func, True)
-        except ImportError:
-            version = ".".join(map(str, sys.version_info[:2]))
-            raise Exception("pywin32 not installed for Python " + version)
-    else:
-        import signal
-        signal.signal(signal.SIGTERM, func)
 
 def main():
 
@@ -63,6 +53,7 @@ def main():
     # Instantiate a QApplication first.
     # Otherwise, Mikibook.create() won't function.
     app = QApplication(sys.argv)
+    print(sys.argv)
 
     # Read notebookList, open the first notebook.
     notebooks = Mikibook.read()
@@ -71,7 +62,8 @@ def main():
         notebooks = Mikibook.read()
 
     if len(notebooks) != 0:
-        if os.path.exists(Mikibook.lockpath):
+        """
+        if os.path.exists(Mikibook.lockpath) and args.command != 'index':
             ret = QMessageBox.question(None, "mikidown - lock file exists", ("It looks like the lock file for "
                 "mikidown already exists. Is mikidown currently running? "
                 "Click no to remove the lock file before rerunning mikidown."), buttons=QMessageBox.Yes|QMessageBox.No)
@@ -83,7 +75,8 @@ def main():
             exit_code = app.exec_()
         else:
             print("Applying single instance per user lock.")
-            os.open(Mikibook.lockpath, os.O_CREAT | os.O_EXCL | os.O_RDWR)
+            lock_fh = os.open(Mikibook.lockpath, os.O_CREAT | os.O_EXCL | os.O_RDWR)
+        """
         settings = Setting(notebooks)
         # Initialize application and main window.
         icon = QIcon(":/icons/mikidown.svg")
@@ -94,16 +87,22 @@ def main():
         tray = MikiTray(icon, window)
         tray.show()
 
+        """
         def cleanup(signum, frame):
             #we need to do this to remove the lock at the end
             app.closeAllWindows()
 
-        set_exit_handler(cleanup)
+        if args.command:
+            signal.signal(signal.SIGTERM, cleanup)
+        """
         exit_code = app.exec_()
 
+        """
         print("Removing single instance per user lock.")
-        if os.path.exists(Mikibook.lockpath):
+        if os.path.exists(Mikibook.lockpath) and args.command != 'index':
+            os.close(lock_fh)
             os.remove(Mikibook.lockpath)
+        """
         sys.exit(exit_code)
 
 if __name__ == '__main__':
